+++ packages/core/src/memory/__tests__/brain-stdp-wave3.test.ts
@@ -340,124 +340,10 @@ describe('T695 — session-bucket pair grouping (real SQLite)', () => {
     await rm(tempDir, { recursive: true, force: true, maxRetries: 5, retryDelay: 500 });
   });
 
-  // Skipped: this is a shared-CI-flaky perf canary, not a behavioral contract.
-  //
-  // The O(n²) complexity guard was originally landed in commit ed81d9fc1 (STDP Wave 3,
-  // T690/T695/T694). The ratio threshold (< 8× slowdown for 4× input) is machine-dependent
-  // and was observed to spike to 10.65× on contended CI, which is why T1093
-  // (commit 9e4db6fb6) marked it `.skip` with a "re-enable once de-flaked" TODO.
-  //
-  // That skip was then accidentally reverted by T1113 (commit 5abf9f486 —
-  // "fix(T1113): add ./code/unfold and ./code/search to @cleocode/nexus exports"),
-  // whose scope was nexus package.json exports. The stdp-wave3 skip-revert was
-  // collateral damage (likely a bad rebase / stale working-copy). Reverting the
-  // revert preserves T1093's de-flake decision.
-  //
-  // Behavioral coverage for T695 session-bucketing is fully provided by T695-2
-  // (within-session pairs always fire), T695-3 (cross-session within window fires),
-  // and T695-4 (cross-session beyond window does NOT fire). The complexity
-  // assertion is orthogonal to correctness.
-  //
-  // This test also does 6 real DB open/migrate/populate cycles + 6 pipeline runs
-  // inside a single `it()` body; under parallel vitest workers on shared CPU this
-  // can exceed the 60 s test timeout for non-algorithmic reasons (GC, I/O contention).
-  //
-  // T1517: re-enable only after (a) rewriting the trials out-of-band
-  // (separate perf suite, sequential run) AND (b) adopting an absolute-time budget
-  // keyed to single-run cost rather than a ratio.
-  it.skip('T695-1: session-bucket O(n²) guard — ratio-based complexity proof (N=50 vs N=200)', async () => {
-    // Complexity proof: if the algorithm is O(n²), scaling input 4× causes ~16× slowdown.
-    // If the algorithm is O(n log n) or better, scaling 4× causes ≤ ~5× slowdown.
-    // We assert ratio < 8 — this disallows quadratic and is machine-independent.
-    // An absolute sanity ceiling of 60 s catches truly broken implementations.
-    //
-    // Each run uses a fresh temp dir so DB state is clean and the two measurements
-    // are independent. The beforeEach tempDir is used for the SMALL run; a second
-    // temp dir is created inline for the LARGE run and cleaned up before the test ends.
-
-    const { closeBrainDb } = await import('../../store/memory-sqlite.js');
-    const { applyStdpPlasticity } = await import('../brain-stdp.js');
-
-    // Helper: populate a fresh DB with numSessions × rowsPerSession retrieval rows
-    // and return the measured duration of a single applyStdpPlasticity call.
-    async function measureRun(
-      dir: string,
-      numSessions: number,
-      rowsPerSession: number,
-    ): Promise<number> {
-      closeBrainDb();
-      const nativeDb = await setupDb(dir);
-
-      nativeDb.exec('BEGIN');
-      const insertStmt = nativeDb.prepare(
-        `INSERT INTO brain_retrieval_log
-           (query, entry_ids, entry_count, source, session_id, reward_signal, created_at)
-         VALUES ('q', ?, 1, 'test', ?, NULL, datetime('now', ?))`,
-      );
-      for (let s = 0; s < numSessions; s++) {
-        const sessionId = `ses_perf_${s}`;
-        // Spread sessions over 0..29 days ago so pairs span the full lookback window
-        const daysBucket = numSessions > 1 ? (s * 29) / (numSessions - 1) : 0;
-        for (let r = 0; r < rowsPerSession; r++) {
-          const entryId = `obs:perf-s${s}-r${r}`;
-          const secondsOffset = Math.round(daysBucket * 86400) + r * 10;
-          insertStmt.run(JSON.stringify([entryId]), sessionId, `-${secondsOffset} seconds`);
-        }
-      }
-      nativeDb.exec('COMMIT');
-
-      closeBrainDb();
-      const startMs = Date.now();
-      await applyStdpPlasticity(dir, {
-        lookbackDays: 30,
-        pairingWindowMs: 24 * 60 * 60 * 1000, // 24h
-      });
-      return Date.now() - startMs;
-    }
-
-    // Take MEDIAN of 3 runs each to smooth out CPU-contention spikes when this
-    // test runs in parallel with other vitest workers. A single run can be
-    // arbitrarily delayed by GC/context-switches; median tracks the true cost.
-    const medianOf = (values: number[]): number => {
-      const sorted = [...values].sort((a, b) => a - b);
-      const mid = Math.floor(sorted.length / 2);
-      return sorted.length % 2 === 0 ? (sorted[mid - 1]! + sorted[mid]!) / 2 : sorted[mid]!;
-    };
-
-    // SMALL run: 5 sessions × 10 rows = 50 spikes — 3 trials
-    const smallDir = tempDir; // managed by beforeEach/afterEach
-    const smallTrials: number[] = [];
-    for (let i = 0; i < 3; i++) smallTrials.push(await measureRun(smallDir, 5, 10));
-    const timeSmallMs = medianOf(smallTrials);
-
-    // LARGE run: 20 sessions × 10 rows = 200 spikes (4× input) — 3 trials
-    const largeDir = await mkdtemp(join(tmpdir(), 'cleo-stdp-wave3-t695-large-'));
-    const largeTrials: number[] = [];
-    try {
-      process.env['CLEO_DIR'] = join(largeDir, '.cleo');
-      for (let i = 0; i < 3; i++) largeTrials.push(await measureRun(largeDir, 20, 10));
-    } finally {
-      closeBrainDb();
-      process.env['CLEO_DIR'] = join(smallDir, '.cleo');
-      await rm(largeDir, { recursive: true, force: true, maxRetries: 5, retryDelay: 500 });
-    }
-    const timeLargeMs = medianOf(largeTrials);
-
-    // Sanity: median runs must complete within 60 s individually (catches totally broken impl)
-    expect(timeSmallMs).toBeLessThan(60_000);
-    expect(timeLargeMs).toBeLessThan(60_000);
-
-    // Complexity assertion: 4× input must not cause more than 8× slowdown.
-    // Linear O(n) → ratio ~4. Log-linear O(n log n) → ratio ~5. Quadratic O(n²) → ratio ~16.
-    // A ratio < 8 proves the implementation is sub-quadratic on this dataset.
-    // Guard against division-by-zero on extremely fast machines (< 1 ms small run).
-    const smallFloor = Math.max(timeSmallMs, 1);
-    const ratio = timeLargeMs / smallFloor;
-    expect(ratio).toBeLessThan(8);
-
-    // Results sanity: both runs must return valid shape
-    // (verified implicitly — measureRun would throw on malformed result)
-  });
+  // T695-1 (O(n²) complexity canary) was moved to brain-stdp-wave3.perf.test.ts
+  // (gated by RUN_PERF=1) because it runs 6 DB cycles × 3 trials under vitest
+  // parallel workers and flaked on contended CI (T1093/T1517). Behavioral coverage
+  // for T695 session-bucketing is provided by T695-2/T695-3/T695-4 below.
 
   it('T695-2: within-session pairs always generated regardless of session size', async () => {
     const nativeDb = await setupDb(tempDir);
